реф: разбираю проект на папки, разбираю компонент конструктор
