+++ src/components/burger-constructor/ingredient-detail-modal/style.module.scss
