+++ src/components/burger-constructor/style.module.scss
@@ -1,7 +1,3 @@
 .container {
-  width: 100%;
-  height: 100px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
+  width: 97%;
 }
