+++ src/components/burger-ingredients/style.module.scss
@@ -90,10 +90,6 @@
   text-align: center;
 }
 
-.item-in-stack {
-  width: 97%;
-}
-
 .order {
   height: 64px;
   width: 100%;
